Update reservations controller

+++ Backend/Backend/Models/RoomDto.cs
@@ -5,7 +5,7 @@
         public required int Capacity { get; set; }
         public required string Name { get; set; }
         public required string Description { get; set; }
-        public required int Price { get; set; }
+        public required float Price { get; set; }
         public required int HotelId { get; set; }
     }
 }
